create handleClickEvents function
Most of the code inside handleClickEvents function is not exactly new. Previously, they were inside
of an anonymous function that was called in the onClick event handler but now they are
inside of a named function.The only notable change is,now the function checks if the
game has a winner or not with every click.