added new imports
added readymade human and computer objects as imports. This is to
ensure that I can use the objects wherever I want in the index.js. Since
I try to write everything inside functions, previously I had to pass
human and computer objects as parameters to the functions and that created a problem
when I tried to add handleCLickEvents function as click event handler which
I could not able to solve without this new approach.