add function to remove all event listener in ai gameboard
abortController api is more convenient than writing a function with removeEventListener
to remove all event listener in ai gameboard,but it does not work in this codebase because
we added 'once:true' to the event listeners and the event listeners are removed after they
are triggered. AbortController only works when all event listeners are intact.